refactor(visualization): reorder folders
