--- libs/ui-react-visualization/src/chart/index.ts
@@ -1,2 +1,0 @@
-export * from './line';
-export * from './internals';
