+++ libs/ui-react-visualization/src/index.ts
@@ -1,1 +1,2 @@
-export * from './chart';
+export * from './lib/Components';
+export * from './utils';
